pyproject.toml
[project]
name = "adk-fleet-safety-multi-agent-system"
version = "0.1.0"
description = "Multi-agent system combining real-time safety monitoring, intelligent EV-aware route planning and predictive analytics to cut incidents"
authors = [
    {name = "Etty Sekhon", email = "etty.sekhon@email.com"},
]
dependencies = [
    "google-adk[eval]>=1.15.0,<2.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[evaluation,agent-engines]>=1.118.0,<2.0.0",
    "protobuf>=6.31.1,<7.0.0",
    # Fleet Safety Specific Dependencies
    "a2a-sdk>=0.3.16",
    "google-maps-mcp-server",
    "httpx>=0.27.0",
    "polyline>=2.0.3",
    "python-a2a>=0.5.10",
    "python-dotenv>=1.2.1",
    "mcp>=1.22.0",
    "click>=8.3.0",
]
requires-python = ">=3.10,<3.14"

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
    "black>=25.11.0",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "locust>=2.42.6",
]

[project.optional-dependencies]
jupyter = [
    "jupyter>=1.0.0,<2.0.0",
]
lint = [
    "ruff>=0.4.6,<1.0.0",
    "mypy>=1.15.0,<2.0.0",
    "codespell>=2.2.0,<3.0.0",
    "types-pyyaml>=6.0.12.20240917,<7.0.0",
    "types-requests>=2.32.0.20240914,<3.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E", "F", "W", "I", "C", "B", "UP", "RUF", "N", "SIM", "TID"
]
ignore = ["E501", "C901", "B006", "W292"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
google-maps-mcp-server = { git = "https://github.com/ettysekhon/google-maps-mcp-server.git" }
